requirements.txt
Flask
sentence-transformers
pandas
numpy
pinecone-client
python-dotenv
requests